requirements.txt
transformers
peft
sentence-transformers
faiss-cpu
pandas
fastapi
uvicorn
torch
